book_store/requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.7
django-crispy-forms==2.3
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
